Added checks for targets and dependencies.
