+++ .devcontainer/Dockerfile
@@ -1,9 +1,11 @@
-FROM rustlang/rust:nightly
+FROM ubuntu:latest
 ENV DEBIAN_FRONTEND=noninteractive
 RUN apt-get update && apt-get install -y \
     build-essential \
     cmake \
+    ninja-build \ 
     clang \
+    curl \ 
     vim \
     pkg-config \ 
     libssl-dev \
