+++ .devcontainer/devcontainer.json
@@ -19,4 +19,4 @@
 			]
 		}
 	}
-} 
+}
