+++ bsan-script/src/main.rs
@@ -1,13 +1,14 @@
-// Our build script is a modified copy of Miri's build script.
+// Our build script combines many preexisting components from Miri's build script
+// and the Rust compiler's bootstrap script. 
 #![feature(io_error_more)]
 use anyhow::Result;
 use clap::{command, Parser};
 use commands::Component;
 //mod commands;
 mod commands;
-mod download;
 mod env;
 mod utils;
+mod setup;
 
 static TOOLCHAIN_NAME: &str = "bsan";
