+++ bsan-script/src/utils.rs
@@ -1,14 +1,15 @@
-use std::fs::canonicalize;
+use std::fs::{self, canonicalize, File};
 use std::io;
-use std::io::Write;
-use std::path::PathBuf;
+use std::io::{BufReader, Write};
+use std::path::{Path, PathBuf};
 
 use anyhow::{anyhow, Context, Result};
 use path_macro::path;
 use rustc_version::VersionMeta;
 use strum::IntoEnumIterator;
 use strum_macros::EnumIter;
-use xshell::{cmd, Shell};
+use xshell::{cmd, Cmd, Shell};
+use xz2::bufread::XzDecoder;
 
 pub fn show_error_(msg: &impl std::fmt::Display) -> ! {
     eprintln!("fatal error: {msg}");
@@ -150,9 +151,169 @@ pub fn install_git_hooks(root_dir: &PathBuf) -> Result<()> {
         {
             continue;
         }
-
         install_hook(hook);
     }
+    Ok(())
+}
 
+static CURL_FLAGS: &[&str] = &[
+    // follow redirect
+    "--location",
+    // timeout if speed is < 10 bytes/sec for > 30 seconds
+    "--speed-time",
+    "30",
+    "--speed-limit",
+    "10",
+    // timeout if cannot connect within 30 seconds
+    "--connect-timeout",
+    "30",
+    // if there is an error, don't restart the download,
+    // instead continue where it left off.
+    "--continue-at",
+    "-",
+    // retry up to 3 times.  note that this means a maximum of 4
+    // attempts will be made, since the first attempt isn't a *re*try.
+    "--retry",
+    "3",
+    // show errors, even if --silent is specified
+    "--show-error",
+    // set timestamp of downloaded file to that of the server
+    "--remote-time",
+    // fail on non-ok http status
+    "--fail",
+];
+
+fn extract_curl_version(out: &[u8]) -> semver::Version {
+    let out = String::from_utf8_lossy(out);
+    // The output should look like this: "curl <major>.<minor>.<patch> ..."
+    out.lines()
+        .next()
+        .and_then(|line| line.split(" ").nth(1))
+        .and_then(|version| semver::Version::parse(version).ok())
+        .unwrap_or(semver::Version::new(1, 0, 0))
+}
+
+fn curl_version(sh: &Shell) -> semver::Version {
+    let curl = cmd!(sh, "curl -V");
+    let Ok(out) = curl.output() else { return semver::Version::new(1, 0, 0) };
+    let out = out.stdout;
+    extract_curl_version(&out)
+}
+
+pub fn download_file(
+    sh: &Shell,
+    url: &Path,
+    destination: &Path,
+    help_on_error: &str,
+) -> Result<()> {
+    println!("downloading {}", url.display());
+
+    if destination.exists() {
+        fs::remove_file(destination)?;
+    }
+
+    let mut curl: Cmd<'_> = cmd!(sh, "curl");
+    curl = curl.args([
+        // output file
+        "--output",
+        destination.to_str().unwrap(),
+    ]);
+    curl = curl.args(CURL_FLAGS);
+    curl = curl.arg(url);
+
+    if is_running_on_ci() {
+        curl = curl.arg("--silent");
+    } else {
+        curl = curl.arg("--progress-bar");
+    }
+    // --retry-all-errors was added in 7.71.0, don't use it if curl is old.
+    if curl_version(sh) >= semver::Version::new(7, 71, 0) {
+        curl = curl.arg("--retry-all-errors");
+    }
+
+    if curl.quiet().run().is_err() {
+        if !help_on_error.is_empty() {
+            show_error!("{help_on_error}");
+        }
+        std::process::exit(1);
+    }
+    Ok(())
+}
+
+pub fn unpack(tarball: &Path, dst: &Path, pattern: &str) -> Result<()> {
+    if !dst.exists() {
+        std::fs::create_dir_all(dst)?;
+    }
+    println!("extracting {}", tarball.display());
+    // `tarball` ends with `.tar.xz`; strip that suffix
+    // example: `rust-dev-nightly-x86_64-unknown-linux-gnu`
+    let uncompressed_filename =
+        Path::new(tarball.file_name().expect("missing tarball filename")).file_stem().unwrap();
+    let directory_prefix = Path::new(Path::new(uncompressed_filename).file_stem().unwrap());
+
+    // decompress the file
+    let data = File::open(tarball)?;
+    let decompressor = XzDecoder::new(BufReader::new(data));
+    let mut tar = tar::Archive::new(decompressor);
+
+    for member in tar.entries()? {
+        let mut member = member?;
+        let original_path = member.path()?.into_owned();
+        if original_path == directory_prefix {
+            continue;
+        }
+        let mut short_path = original_path.strip_prefix(directory_prefix)?;
+        short_path = short_path.strip_prefix(pattern).unwrap_or(short_path);
+        let dst_path = dst.join(short_path);
+
+        if !member.unpack_in(dst)? {
+            panic!("path traversal attack ??");
+        }
+        let src_path = dst.join(original_path);
+        if src_path.is_dir() && dst_path.exists() {
+            continue;
+        }
+        move_file(src_path, dst_path)?;
+    }
+    let dst_dir = dst.join(directory_prefix);
+    if dst_dir.exists() {
+        fs::remove_dir_all(&dst_dir)?;
+    }
+    Ok(())
+}
+
+/// Rename a file if from and to are in the same filesystem or
+/// copy and remove the file otherwise
+fn move_file<P: AsRef<Path>, Q: AsRef<Path>>(from: P, to: Q) -> io::Result<()> {
+    match fs::rename(&from, &to) {
+        Err(e) if e.kind() == io::ErrorKind::CrossesDevices => {
+            std::fs::copy(&from, &to)?;
+            std::fs::remove_file(&from)
+        }
+        r => r,
+    }
+}
+
+#[allow(dead_code)]
+pub fn ask_to_run(cmd: Cmd<'_>, ask: bool, text: &str) -> Result<()> {
+    // Disable interactive prompts in CI (GitHub Actions, Travis, AppVeyor, etc).
+    // Azure doesn't set `CI` though (nothing to see here, just Microsoft being Microsoft),
+    // so we also check their `TF_BUILD`.
+    let is_ci = is_running_on_ci();
+    if ask && !is_ci {
+        let mut buf = String::new();
+        print!("I will run `{cmd}` to {text}. Proceed? [Y/n] ");
+        io::stdout().flush().unwrap();
+        io::stdin().read_line(&mut buf).unwrap();
+        match buf.trim().to_lowercase().as_ref() {
+            // Proceed.
+            "" | "y" | "yes" => {}
+            "n" | "no" => show_error!("aborting as per your request"),
+            a => show_error!("invalid answer `{}`", a),
+        };
+    } else {
+        eprintln!("Running `{cmd:?}` to {text}.");
+    }
+    cmd.run()?;
     Ok(())
 }
