--- bsan.toml
@@ -1,2 +1,0 @@
-artifact_url = "https://github.com/BorrowSanitizer/rust/releases/download/"
-tag = "rolling-2025-06-06"
