+++ config.toml
@@ -1,0 +1,12 @@
+artifact_url = "https://github.com/BorrowSanitizer/rust/releases/download/"
+tag = "rolling"
+sha = "2cb886eebe58c9a4948fe3ed9117f895dab8049b"
+dependencies = [
+    "cmake",
+    "ninja",
+    "curl"
+]
+targets = [
+    "x86_64-unknown-linux-gnu",
+    "aarch64-unknown-linux-gnu"
+]
